fix(cli): frontmatter descriptions + extract shared fixture-utils, -22 diff lines
- Add hand-curated description to 11 fixture meta.json files
- Extract FixtureMeta, metaToCompoundMeta, FIXTURES, FIXTURES_DIR into
  shared fixture-utils.ts to eliminate duplication between test + convergence
- Add clarifying comments on description contract in renderers
