+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/device_vector_v5/meta.json
@@ -6,6 +6,7 @@
   "namespace_path": [
     "thrust"
   ],
+  "description": "A dynamically-sized container for device memory with automatic memory management.",
   "golden_page": "thrust/device_vector_v5.mdx",
   "status": "MATCHED"
 }
