+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/group_member_example_v5/meta.json
@@ -7,6 +7,7 @@
     "thrust",
     "mr"
   ],
+  "description": "A memory resource adaptor that pools and caches allocations using a separate bookkeeper for managing memory inaccessible from the host.",
   "golden_page": "thrust/group_member_example_v5.mdx",
   "status": "MATCHED"
 }
