+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/pointer_v5/meta.json
@@ -6,6 +6,7 @@
   "namespace_path": [
     "thrust"
   ],
+  "description": "A tagged pointer type that stores a pointer to an object allocated in memory, generalizing device_ptr with configurable backend systems.",
   "golden_page": "thrust/pointer_v5.mdx",
   "status": "MATCHED"
 }
