+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/raises_example_v5/meta.json
@@ -6,6 +6,7 @@
   "namespace_path": [
     "cuda"
   ],
+  "description": "An owning wrapper for cudaStream_t providing RAII-based stream lifecycle management.",
   "golden_page": "libcudacxx/raises_example_v5.mdx",
   "status": "MATCHED"
 }
