+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/simple_struct_v5/meta.json
@@ -6,6 +6,7 @@
   "namespace_path": [
     "cub"
   ],
+  "description": "Arg max functor that keeps the value and offset of the first occurrence of the larger item.",
   "golden_page": "cub/simple_struct_v5.mdx",
   "status": "MATCHED"
 }
