+++ packages/cli/library-docs-generator/cpp/src/__test__/fixtures/warp_reduce_v5/meta.json
@@ -6,6 +6,7 @@
   "namespace_path": [
     "cub"
   ],
+  "description": "Collective methods for computing parallel reductions across a CUDA thread warp.",
   "golden_page": "cub/warp_reduce_v5.mdx",
   "status": "MATCHED"
 }
